python-ocr/requirements.txt
fastapi
uvicorn
pytesseract
opencv-python-headless
pillow
python-multipart
